barista(search): Fixes an issue with search on overview pages.
The focuspulling of keyInputs on overview pages, which should help you find things faster on
overview pages, was interfering with the input in the search.
To prevent this, we have moved them to the same keyboardEvent (keyup instead of input), and added a
propagation blocker to the search.
Also added an error handler for the rest call to the search, which was failing the subscription.